Move mock benchmark models to test module
